experiment(016): implement Value tagged-union struct — measure impact
Full implementation of proposal 016: replace cbor.Value interface with
a discriminated struct carrying Kind + inline storage for all variants.

All tests pass across all modules (root, cose, cwt, eat, edhoc).
Migration was ~140 call sites, mechanically transformed.

RESULTS (AMD EPYC 7R13, linux/amd64):

  Allocations improved as predicted:
  - DecodeNestedMapStrict: 46 → 19 allocs/op (-59%)
  - DecodeScalars:          2 →  1 alloc/op  (-50%)
  - MarshalSign1:          10 →  5 allocs/op (-50%)

  Throughput REGRESSED severely:
  - EncodeScalars:      615 → 149 MB/s (-76%)
  - EncodeNestedMap:    688 →  96 MB/s (-86%)
  - DecodeScalars:      460 → 132 MB/s (-71%)
  - DecodeNestedMapStrict: 58 → 31 MB/s (-47%)
  - MarshalSign1:       400 → 505 ns/op (-21%)

ROOT CAUSE: The Value struct is 104 bytes. Every recursive
encode/decode call copies it by value. With 1000-element arrays the
aggregate copy bandwidth overwhelms the allocation savings. The
interface model copies only 16 bytes (type ptr + data ptr) per call.

CONCLUSION: Proposal 016 is rejected in its current form. The struct
size must be reduced (unsafe union, ~40 bytes) or the recursive paths
restructured to avoid per-value copies before this approach can work.

This commit preserves the implementation for reference. The next commit
reverts it.

Via: git.commit.create (mcp-server-go-git/v0.1.0-8-gb11cc3f)

+++ edhoc/keyschedule.go
@@ -62,7 +62,7 @@ func edhocKDF(prk, transcriptHash []byte, label string, length int) ([]byte, err
 	info, err := encodeCBORArray(
 		cbor.Bytes(transcriptHash),
 		cbor.Text(label),
-		cbor.Uint(length),
+		cbor.Uint(uint64(length)),
 	)
 	if err != nil {
 		return nil, err
@@ -80,7 +80,7 @@ func encodeExporterInfo(th4 []byte, label string, context []byte, length int) ([
 		cbor.Bytes(th4),
 		cbor.Text(label),
 		cbor.Bytes(context),
-		cbor.Uint(length),
+		cbor.Uint(uint64(length)),
 	)
 }
